Scan to set log level dynamicly

+++ android/sdk/src/main/java/com/taobao/weex/common/IWXBridge.java
@@ -59,7 +59,7 @@ public interface IWXBridge extends IWXObject {
   void updateInitFrameworkParams(String key, String value, String desc);
 
 
-  void setLogType(int type, boolean isPerf);
+  void setLogType(float type, boolean isPerf);
 
   void refreshInstance(String instanceId, String namespace, String function, WXJSObject[] args);
